requirements.txt
django
django-ninja
pillow
ipython
httpx
celery
redis